refactor : inputView 와 resultView 로 분리하기
