+++ README.md
@@ -42,7 +42,7 @@
 - 외부에서 로또 번호들을 볼 수 있다.
 - 외부로부터 당첨번호와 보너스번호를 입력받으면 몇등에 당첨되었는지 반환해준다.
 
-### lotto.view.LottoAutoView 의 역할
+### lotto.view.LottoAutoInputView 의 역할
 - 사용자로부터 로또구매금액을 입력받는다.
 - 사용자로부터 당첨번호와 보너스번호를 입력받는다.
   
@@ -76,7 +76,7 @@
 - 외부에서 로또 번호들을 볼 수 있다.
 - 외부로부터 당첨번호와 보너스번호를 입력받으면 몇등에 당첨되었는지 반환해준다.
 
-### lotto.view.LottoManualView 의 역할
+### lotto.view.LottoManualInputView 의 역할
 - 사용자로부터 로또구매금액을 입력받는다.
 - 사용자로부터 로또수동갯수를 입력받는다.
 - 사용자로부터 로또수동번호들을 입력받는다.
