+++ src/main/java/lotto/view/LottoAutoInputView.java
@@ -1,0 +1,32 @@
+package lotto.view;
+
+import lotto.model.WinningResult;
+import lotto.model.LottoModel;
+
+import java.util.List;
+import java.util.Scanner;
+
+public class LottoAutoInputView {
+
+    public int inputPrice() {
+        Scanner scanner = new Scanner(System.in);
+
+        System.out.println("구입금액을 입력해 주세요.");
+        return scanner.nextInt();
+    }
+
+    public String inputWinningNumbers() {
+        Scanner scanner = new Scanner(System.in);
+
+        System.out.println();
+        System.out.println("지난 주 당첨 번호를 입력해 주세요.");
+        return scanner.nextLine();
+    }
+
+    public String inputBonusNumber() {
+        Scanner scanner = new Scanner(System.in);
+
+        System.out.println("보너스 볼을 입력해 주세요.");
+        return scanner.nextLine();
+    }
+}
